+++ source/extensions/lightspeed.trex.sidebar/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.2.1]
+### Changed
+- Updated WorkspaceWidget interface to call super().show() for proper visibility tracking
+
 ## [1.2.0]
 ### Added
 - Added disabled_tooltip support for sidebar items